+++ security/sample/src/main/java/ru/surfstudio/android/security/sample/app/CustomApp.kt
@@ -3,7 +3,7 @@ package ru.surfstudio.android.security.sample.app
 import androidx.multidex.MultiDexApplication
 import ru.surfstudio.android.activity.holder.ActiveActivityHolder
 import ru.surfstudio.android.sample.dagger.app.dagger.DefaultAppModule
-import ru.surfstudio.android.security.app.AppDebuggableChecker
+import ru.surfstudio.android.security.app.ReleaseAppChecker
 import ru.surfstudio.android.security.sample.BuildConfig
 import ru.surfstudio.android.security.sample.app.dagger.CustomAppComponent
 import ru.surfstudio.android.security.sample.app.dagger.DaggerCustomAppComponent
@@ -20,7 +20,9 @@ class CustomApp : MultiDexApplication() {
 
     override fun onCreate() {
         super.onCreate()
-        AppDebuggableChecker.check(this, BuildConfig.CHECK_DEBUGGABLE)
+        if (BuildConfig.CHECK_DEBUGGABLE) {
+            ReleaseAppChecker.checkReleaseApp(this)
+        }
         initInjector()
         registerSessionManager()
     }